requirements.txt
moshi>=0.2.6
pdfminer.six
torch